+++ Tests/MinimalEndpoints.Tests/Usings.cs
@@ -1,3 +1,4 @@
-global using Microsoft.VisualStudio.TestTools.UnitTesting;
 global using FluentAssertions;
 global using Genesis;
+global using Genesis.Http;
+global using Microsoft.VisualStudio.TestTools.UnitTesting;
